+++ src/collective/geo/faceted/widget/interfaces.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+from eea.facetednavigation.interfaces import IWidget
+
+
+class IBoundsWidget(IWidget):
+    """ Results Filter widget
+    """